+++ MailSender/RecipientView.xaml
