+++ MailSender/RecipientView.xaml.cs
