+++ MailSender/RecipientsView.xaml
