+++ MailSender/RecipientsView.xaml.cs
