+++ MailSender/ViewModel/ViewModelLocator.cs
@@ -1,0 +1,31 @@
+using GalaSoft.MvvmLight;
+using GalaSoft.MvvmLight.Ioc;
+using CommonServiceLocator;
+using MailSender.ViewModels;
+
+namespace MailSender.ViewModel
+{
+    /// <summary>
+    /// This class contains static references to all the view models in the
+    /// application and provides an entry point for the bindings.
+    /// </summary>
+    public class ViewModelLocator
+    {
+        /// <summary>
+        /// Initializes a new instance of the ViewModelLocator class.
+        /// </summary>
+        public ViewModelLocator()
+        {
+            ServiceLocator.SetLocatorProvider(() => SimpleIoc.Default);
+
+            SimpleIoc.Default.Register<MainWindowViewModel>();
+        }
+
+        public MainWindowViewModel MainWindowViewModel => ServiceLocator.Current.GetInstance<MainWindowViewModel>();
+
+        public static void Cleanup()
+        {
+            // TODO Clear the ViewModels
+        }
+    }
+}
